third_party/blink/web_tests/fast/forms/ValidityState-stepMismatch-expected.txt

CONSOLE WARNING: The specified value "2" does not conform to the required format.  The format is "#rrggbb" where rr, gg, bb are two-digit hexadecimal numbers.
Check stepMismatch results for unsupported types.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Unsupported types
PASS input.type = "text"; input.step = "3"; input.min = ""; input.value = "2"; input.validity.stepMismatch is false
PASS input.type = "button"; input.validity.stepMismatch is false
PASS input.type = "checkbox"; input.validity.stepMismatch is false
PASS input.type = "color"; input.validity.stepMismatch is false
PASS input.type = "email"; input.validity.stepMismatch is false
PASS input.type = "hidden"; input.validity.stepMismatch is false
PASS input.type = "image"; input.validity.stepMismatch is false
PASS input.type = "khtml_isindex"; input.validity.stepMismatch is false
PASS input.type = "passwd"; input.validity.stepMismatch is false
PASS input.type = "radio"; input.validity.stepMismatch is false
PASS input.type = "reset"; input.validity.stepMismatch is false
PASS input.type = "search"; input.validity.stepMismatch is false
PASS input.type = "submit"; input.validity.stepMismatch is false
PASS input.type = "tel"; input.validity.stepMismatch is false
PASS input.type = "url"; input.validity.stepMismatch is false
PASS input.type = "file"; input.validity.stepMismatch is false
PASS successfullyParsed is true

TEST COMPLETE
